Djirgo/requirements.txt
asgiref==3.7.2
Django==3.2.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework==3.12.4